refactor: parser 테스트 클래스 생성
- 기존 ValidationTest 클래스에 작성된 파서 관련 테스트 함수를 ParserTest 클래스로 분리
- 테스트 코드의 책임을 명확히 분리하고 가독성을 향상시킴
